requirements.txt
PyQt6>=6.6.0
arxiv>=2.1.0
google-genai>=1.0.0
openai>=1.50.0
ollama>=0.4.0
